Used react router to generate /notes and /todos routes
