+++ src/components/App.js
@@ -1,17 +1,12 @@
 import React from "react";
-import { values } from "lodash";
-import TodoList from "./TodoList";
-import UserPreferences from "./UserPreferences";
-import TodoListCreationForm from "./TodoListCreationForm";
-import Button from "./library/Button";
 import { connect } from "react-redux";
+
+import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
+
+import Button from "./library/Button";
 import NotificationCenter from "./NotificationCenter";
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-  Link
-} from "react-router-dom";
+import Notes from "./pages/Notes";
+import Todos from "./pages/Todos";
 
 // Small comment to open the PR
 class App extends React.Component {
@@ -19,6 +14,7 @@ class App extends React.Component {
     return (
       <Router>
         <div className="App">
+
           <h1 className="main-title">Scotto TodoList App !</h1>
           <Button
             additionalClassName="save-data"
@@ -44,53 +40,31 @@ class App extends React.Component {
           </nav>
 
           <div className="app-container">
-            <div className="bloc-1">
-              <UserPreferences/>
-              <TodoListCreationForm onCreate={this.props.addTodoList}/>
-            </div>
-            <div className="all-todoLists-container">
-              {this.props.todoLists.map((todoList) => {
-                return (
-                  <TodoList
-                    key={todoList.id}
-                    name={todoList.name}
-                    onDelete={this.props.deleteTodoList}
-                    id={todoList.id}
-                  />
-                );
-              })}
-            </div>
+            <Switch>
+              <Route path="/notes">
+                <Notes />
+              </Route>
+              <Route path="/todos">
+                <Todos />
+              </Route>
+              <Route path="/">
+                <Todos />
+              </Route>
+            </Switch>
           </div>
+
         </div>
       </Router>
     );
   }
 }
 
-const mapStateToProps = (state, props) => {
-  return {
-    todoLists: values(state.todoLists)
-  }
-}
-
 const mapDispatchToProps = (dispatch, props) => {
   return {
-    addTodoList: (name) => dispatch({
-      type: 'ADD_TODOLIST',
-      payload: {
-        name
-      }
-    }),
-    deleteTodoList: (todoListId) => dispatch({
-      type: 'DELETE_TODOLIST',
-      payload: {
-        todoListId
-      }
-    }),
     saveData: () => dispatch({
       type: 'SAVE_DATA'
     })
   }
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(App);
+export default connect(null, mapDispatchToProps)(App);
